+++ common/src/main/java/dev/ryanhcode/sable/physics/floating_block/FloatingBlockController.java
@@ -27,21 +27,20 @@ public class FloatingBlockController {
     private static final Vector3d localLinearVelocity = new Vector3d();
     private static final Vector3d localAngularVelocity = new Vector3d();
     private final FloatingClusterContainer sublevelContainer = new FloatingClusterContainer();
-    List<FloatingClusterContainer> containers = new ArrayList<>();
+    private final List<FloatingClusterContainer> containers = new ArrayList<>();
     private final ServerSubLevel subLevel;
     private final Vector3d previousCenterOfMass = new Vector3d();
-
 
     public FloatingBlockController(final ServerSubLevel subLevel) {
         this.subLevel = subLevel;
     }
 
     public void physicsTick(final double partialPhysicsTick, final double timeStep, final Vector3dc linearVelocity, final Vector3dc angularVelocity, final Vector3d linearImpulse, final Vector3d angularImpulse) {
+        this.containers.clear();
+        this.containers.add(this.sublevelContainer);
 
-        containers.clear();
-        containers.add(sublevelContainer);
-        for(KinematicContraption contraption : subLevel.getPlot().getContraptions()) {
-            FloatingClusterContainer container = contraption.sable$getFloatingClusterContainer();
+        for (final KinematicContraption contraption : this.subLevel.getPlot().getContraptions()) {
+            final FloatingClusterContainer container = contraption.sable$getFloatingClusterContainer();
 
             final Vector3dc lastPosition = new Vector3d(contraption.sable$getPosition(partialPhysicsTick - 1.0f));
             final Quaterniondc lastOrientation = new Quaterniond(contraption.sable$getOrientation(partialPhysicsTick - 1.0f));
@@ -52,18 +51,12 @@ public class FloatingBlockController {
             container.positionOffset.sub(lastPosition, container.velocity);
             SableMathUtils.getAngularVelocity(lastOrientation, container.rotationOffset, container.angularVelocity);
 
-            //System.out.println(lastOrientation.z()+"  "+lastOrientation.w());
-            //System.out.println(container.rotationOffset.z()+"  "+container.rotationOffset.w());
-            //System.out.println(container.angularVelocity.z);
-
             container.velocity.mul(20.0);
             container.angularVelocity.mul(20.0);
-            //container.rotationOffset.transformInverse(container.velocity);
-            //container.rotationOffset.transformInverse(container.angularVelocity);
 
-            container.positionOffset.sub(subLevel.getMassTracker().getCenterOfMass());
+            container.positionOffset.sub(this.subLevel.getMassTracker().getCenterOfMass());
 
-            containers.add(container);
+            this.containers.add(container);
         }
 
 
@@ -71,7 +64,7 @@ public class FloatingBlockController {
 
         localGravity.set(DimensionPhysicsData.getGravity(this.subLevel.getLevel(), this.subLevel.logicalPose().position()));
         this.subLevel.logicalPose().orientation().transformInverse(localGravity);
-        if(!needsTicking())
+        if (!this.needsTicking())
             return;
 
         this.subLevel.logicalPose().orientation().transformInverse(linearVelocity, localLinearVelocity);
@@ -83,16 +76,16 @@ public class FloatingBlockController {
         final QueuedForceGroup dragGroup = this.subLevel.getOrCreateQueuedForceGroup(ForceGroups.DRAG.get());
         final List<Vector3d> recordedFrictionForces = new ObjectArrayList<>();
 
-        for (FloatingClusterContainer container : containers) {
+        for (final FloatingClusterContainer container : this.containers) {
 
             for (final FloatingBlockCluster cluster : container.clusters) {
                 if (cluster.getMaterial().scaleWithPressure())
-                    cluster.getBlockData().computePressureScale(subLevel);
+                    cluster.getBlockData().computePressureScale(this.subLevel);
 
-                this.applyFriction(container,cluster, localGravity, localLinearVelocity, localAngularVelocity, clusterFrictionForce, clusterFrictionTorque);
+                this.applyFriction(container, cluster, localGravity, localLinearVelocity, localAngularVelocity, clusterFrictionForce, clusterFrictionTorque);
 
                 final Vector3d recordedClusterFrictionForce = new Vector3d(clusterFrictionForce);
-                this.recordForce(container,cluster, dragGroup, recordedClusterFrictionForce);
+                this.recordForce(container, cluster, dragGroup, recordedClusterFrictionForce);
                 recordedFrictionForces.add(recordedClusterFrictionForce);
 
                 frictionForce.add(clusterFrictionForce);
@@ -120,19 +113,18 @@ public class FloatingBlockController {
         for (final Vector3d force : recordedFrictionForces) {
             force.mul(timeStep);//forceScale *
         }
-        if(localGravity.lengthSquared()>0)
+        if (localGravity.lengthSquared() > 0)
             this.applyLift(localGravity, linearImpulse, angularImpulse, timeStep);
 
         linearImpulse.fma(timeStep, frictionForce);
         angularImpulse.fma(timeStep, frictionTorque);
     }
 
-    public boolean needsTicking()
-    {
-        if(sublevelContainer.needsTicking())
+    public boolean needsTicking() {
+        if (this.sublevelContainer.needsTicking())
             return true;
-        for (FloatingClusterContainer container : containers) {
-            if(container.needsTicking())
+        for (final FloatingClusterContainer container : this.containers) {
+            if (container.needsTicking())
                 return true;
         }
         return false;
@@ -141,10 +133,10 @@ public class FloatingBlockController {
     private void processBlockChanges() {
 
         this.previousCenterOfMass.sub(this.subLevel.getMassTracker().getCenterOfMass());
-        for (final FloatingBlockCluster cluster : sublevelContainer.clusters) {
+        for (final FloatingBlockCluster cluster : this.sublevelContainer.clusters) {
             cluster.getBlockData().translateOrigin(this.previousCenterOfMass);
         }
-        sublevelContainer.processBlockChanges(this.subLevel.getMassTracker().getCenterOfMass());
+        this.sublevelContainer.processBlockChanges(this.subLevel.getMassTracker().getCenterOfMass());
         this.previousCenterOfMass.set(this.subLevel.getMassTracker().getCenterOfMass());
 
     }
@@ -166,7 +158,7 @@ public class FloatingBlockController {
 
         totalWeightedForce.set(0);
 
-        for (FloatingClusterContainer container : containers) {
+        for (final FloatingClusterContainer container : this.containers) {
 
             for (final FloatingBlockCluster cluster : container.clusters) {
                 final FloatingBlockMaterial material = cluster.getMaterial();
@@ -179,9 +171,9 @@ public class FloatingBlockController {
                     clusterForce *= cluster.getBlockData().getPressureScale();
 
                 //unit: strength * weight
-                double weightedForce = clusterForce * cluster.getBlockData().totalScale;
+                final double weightedForce = clusterForce * cluster.getBlockData().totalScale;
 
-                getTrueWeightedClusterPosition(container,cluster,weightedPositionTemp);
+                this.getTrueWeightedClusterPosition(container, cluster, weightedPositionTemp);
 
                 if (material.preventSelfLift()) {
                     totalForce += weightedForce;
@@ -193,7 +185,7 @@ public class FloatingBlockController {
                     if (this.subLevel.isTrackingIndividualQueuedForces()) {
                         final QueuedForceGroup levitationGroup = this.subLevel.getOrCreateQueuedForceGroup(ForceGroups.LEVITATION.get());
 
-                        this.recordForce(container,cluster, levitationGroup, new Vector3d(localGravity).mul(-weightedForce * timeStep));
+                        this.recordForce(container, cluster, levitationGroup, new Vector3d(localGravity).mul(-weightedForce * timeStep));
                     }
 
                     localGravity.cross(weightedPositionTemp, torqueTemp);//torqueTemp unit: weight * position * gravity
@@ -225,7 +217,7 @@ public class FloatingBlockController {
         if (this.subLevel.isTrackingIndividualQueuedForces()) {
             final QueuedForceGroup levitationGroup = this.subLevel.getOrCreateQueuedForceGroup(ForceGroups.LEVITATION.get());
 
-            for (FloatingClusterContainer container : containers) {
+            for (final FloatingClusterContainer container : this.containers) {
 
                 for (final FloatingBlockCluster cluster : container.clusters) {
                     final FloatingBlockMaterial material = cluster.getMaterial();
@@ -233,7 +225,7 @@ public class FloatingBlockController {
                     final Vector3d force = new Vector3d(localGravity).mul(timeStep * -cluster.getBlockData().totalScale * material.liftStrength());
                     force.mul(scaleFactor);
 
-                    this.recordForce(container,cluster, levitationGroup, force);
+                    this.recordForce(container, cluster, levitationGroup, force);
                 }
             }
         }
@@ -244,13 +236,13 @@ public class FloatingBlockController {
         angularImpulse.fma(timeStep, liftingTorque);
     }
 
-    private void recordForce(final FloatingClusterContainer container,final FloatingBlockCluster cluster, final QueuedForceGroup forceGroup, final Vector3d force) {
-        forceGroup.recordPointForce(getTrueWeightedClusterPosition(container,cluster,new Vector3d()).div(cluster.getBlockData().totalScale).add(this.subLevel.getMassTracker().getCenterOfMass()), force);
+    private void recordForce(final FloatingClusterContainer container, final FloatingBlockCluster cluster, final QueuedForceGroup forceGroup, final Vector3d force) {
+        forceGroup.recordPointForce(this.getTrueWeightedClusterPosition(container, cluster, new Vector3d()).div(cluster.getBlockData().totalScale).add(this.subLevel.getMassTracker().getCenterOfMass()), force);
     }
-    private Vector3d getTrueWeightedClusterPosition(final FloatingClusterContainer container,final FloatingBlockCluster cluster,final Vector3d pos)
-    {
-        container.rotationOffset.transform(cluster.getBlockData().weightedPosition,pos);
-        return pos.fma(cluster.getBlockData().totalScale,container.positionOffset);
+
+    private Vector3d getTrueWeightedClusterPosition(final FloatingClusterContainer container, final FloatingBlockCluster cluster, final Vector3d pos) {
+        container.rotationOffset.transform(cluster.getBlockData().weightedPosition, pos);
+        return pos.fma(cluster.getBlockData().totalScale, container.positionOffset);
     }
 
     private static final Matrix3d containerRotation = new Matrix3d();
@@ -268,10 +260,10 @@ public class FloatingBlockController {
     private static final Vector3d shiftedCenter = new Vector3d();
     private static final Vector3d linearSlowDrag = new Vector3d();
 
-    private void applyFriction(final FloatingClusterContainer container,final FloatingBlockCluster cluster, final Vector3dc localGravity, final Vector3dc linearVelocity, final Vector3dc angularVelocity, final Vector3d frictionForce, final Vector3d frictionTorque) {
+    private void applyFriction(final FloatingClusterContainer container, final FloatingBlockCluster cluster, final Vector3dc localGravity, final Vector3dc linearVelocity, final Vector3dc angularVelocity, final Vector3d frictionForce, final Vector3d frictionTorque) {
 
         double frictionScale = 1;
-        if(cluster.getMaterial().scaleWithGravity())
+        if (cluster.getMaterial().scaleWithGravity())
             frictionScale = localGravity.length();
         if (cluster.getMaterial().scaleWithPressure())
             frictionScale *= cluster.getBlockData().getPressureScale();
@@ -281,7 +273,7 @@ public class FloatingBlockController {
             speedScale = 0;
 
         totalAngularVelocity.set(angularVelocity).add(container.angularVelocity);
-        getTrueWeightedClusterPosition(container,cluster,clusterCenter).div(cluster.getBlockData().totalScale);
+        this.getTrueWeightedClusterPosition(container, cluster, clusterCenter).div(cluster.getBlockData().totalScale);
 
         cluster.getBlockData().outerProduct.scale(1 / cluster.getBlockData().totalScale, averagePositionMatrix);
 
@@ -298,8 +290,8 @@ public class FloatingBlockController {
 
         //velocity of the center of lift in local space
         angularVelocity.cross(clusterCenter, meanVelocity);
-        container.rotationOffset.transform(cluster.getBlockData().weightedPosition,rotatedPos).div(cluster.getBlockData().totalScale);
-        Vector3d extraContainerVelocity = container.angularVelocity.cross(rotatedPos,rotatedPos);
+        container.rotationOffset.transform(cluster.getBlockData().weightedPosition, rotatedPos).div(cluster.getBlockData().totalScale);
+        final Vector3d extraContainerVelocity = container.angularVelocity.cross(rotatedPos, rotatedPos);
         meanVelocity.add(linearVelocity).add(container.velocity).add(extraContainerVelocity);
 
         //center of the shifted position distribution relative to clusterCenter, variance is shiftedPositionMatrix
@@ -350,7 +342,7 @@ public class FloatingBlockController {
     }
 
     private Matrix3d getGravityMatrix(final Vector3dc g, final double verticalDrag, final double horizontalDrag, final Matrix3d target) {
-        if(g.lengthSquared() > 0.00001)
+        if (g.lengthSquared() > 0.00001)
             SableMathUtils.setOuterProduct(g, g, (horizontalDrag - verticalDrag) / g.dot(g), target);
         else
             target.identity();
@@ -372,31 +364,31 @@ public class FloatingBlockController {
         return v * (1 - Math.exp(-k / v)) / k;
     }
 
-    private double getKineticClampingFactor(final Vector3dc currentLinearVelocity,final Vector3dc currentAngularVelocity,final Vector3d frictionForce,final Vector3d frictionTorque,final double timestep) {
+    private double getKineticClampingFactor(final Vector3dc currentLinearVelocity, final Vector3dc currentAngularVelocity, final Vector3d frictionForce, final Vector3d frictionTorque, final double timestep) {
 
-        double numerator = currentLinearVelocity.dot(frictionForce) + currentAngularVelocity.dot(frictionTorque);
-        double denominator = frictionForce.dot(frictionForce)*subLevel.getMassTracker().getInverseMass() +
-                SableMathUtils.multiplyInnerProduct(frictionTorque,subLevel.getMassTracker().getInverseInertiaTensor(),frictionTorque);
-        denominator*=timestep;
-        if(denominator < 1E-10)
+        final double numerator = currentLinearVelocity.dot(frictionForce) + currentAngularVelocity.dot(frictionTorque);
+        double denominator = frictionForce.dot(frictionForce) * this.subLevel.getMassTracker().getInverseMass() +
+                SableMathUtils.multiplyInnerProduct(frictionTorque, this.subLevel.getMassTracker().getInverseInertiaTensor(), frictionTorque);
+        denominator *= timestep;
+        if (denominator < 1E-10)
             return 1;
-        double t = -numerator/denominator;
-        return Math.max(Math.min(t,1),0);
+        final double t = -numerator / denominator;
+        return Math.max(Math.min(t, 1), 0);
     }
 
     public void addFloatingBlock(final BlockState state, final Vector3d pos) {
-        sublevelContainer.addFloatingBlock(state,pos);
+        this.sublevelContainer.addFloatingBlock(state, pos);
     }
 
     public void removeFloatingBlock(final BlockState state, final Vector3d pos) {
-        sublevelContainer.removeFloatingBlock(state,pos);
+        this.sublevelContainer.removeFloatingBlock(state, pos);
     }
 
     public void queueAddFloatingBlock(final BlockState state, final BlockPos pos) {
-        sublevelContainer.queueAddFloatingBlock(state,pos);
+        this.sublevelContainer.queueAddFloatingBlock(state, pos);
     }
 
     public void queueRemoveFloatingBlock(final BlockState state, final BlockPos pos) {
-        sublevelContainer.queueRemoveFloatingBlock(state,pos);
+        this.sublevelContainer.queueRemoveFloatingBlock(state, pos);
     }
 }