+++ common/src/main/java/dev/ryanhcode/sable/sublevel/storage/serialization/SubLevelStorage.java
@@ -1,12 +1,15 @@
 package dev.ryanhcode.sable.sublevel.storage.serialization;
 
 import dev.ryanhcode.sable.Sable;
+import dev.ryanhcode.sable.SableServerConfig;
+import dev.ryanhcode.sable.companion.math.BoundingBox3d;
 import dev.ryanhcode.sable.sublevel.storage.holding.GlobalSavedSubLevelPointer;
 import dev.ryanhcode.sable.sublevel.storage.holding.SavedSubLevelPointer;
 import dev.ryanhcode.sable.sublevel.storage.holding.SubLevelHoldingChunk;
 import dev.ryanhcode.sable.sublevel.storage.region.SubLevelRegionFile;
 import dev.ryanhcode.sable.sublevel.storage.region.SubLevelStorageFile;
 import it.unimi.dsi.fastutil.longs.Long2ObjectLinkedOpenHashMap;
+import it.unimi.dsi.fastutil.objects.ObjectIterator;
 import net.minecraft.FileUtil;
 import net.minecraft.core.SectionPos;
 import net.minecraft.nbt.CompoundTag;
@@ -14,6 +17,7 @@ import net.minecraft.util.ExceptionCollector;
 import net.minecraft.world.level.ChunkPos;
 import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.NotNull;
+import org.joml.Vector3d;
 
 import java.io.IOException;
 import java.nio.file.Path;
@@ -49,7 +53,12 @@ public class SubLevelStorage implements AutoCloseable {
         }
 
         if (this.regionCache.size() >= MAX_CACHE_SIZE) {
-            this.regionCache.removeLast().close();
+            final SubLevelRegionFile last = this.regionCache.removeLast();
+            if (last.isEmpty() && SableServerConfig.SUB_LEVEL_STORAGE_PRUNING.get()) {
+                last.delete();
+            } else {
+                last.close();
+            }
         }
 
         final Path path = this.getPath(chunkPos);
@@ -68,7 +77,13 @@ public class SubLevelStorage implements AutoCloseable {
         }
 
         if (this.storageCache.size() >= MAX_CACHE_SIZE) {
-            this.storageCache.removeLast().close();
+            final SubLevelStorageFile last = this.storageCache.removeLast();
+
+            if (last.isEmpty() && SableServerConfig.SUB_LEVEL_STORAGE_PRUNING.get()) {
+                last.delete();
+            } else {
+                last.close();
+            }
         }
 
         FileUtil.createDirectoriesSafe(this.folder);
@@ -98,7 +113,15 @@ public class SubLevelStorage implements AutoCloseable {
             regionFile.trySave(chunkPos.getRegionLocalX(), chunkPos.getRegionLocalZ(), holdingChunk);
         } catch (final IOException e) {
             Sable.LOGGER.error("Failed to save holding chunk for {}", chunkPos, e);
+        }
+    }
 
+    public void attemptRemoveHoldingChunk(final ChunkPos chunkPos) {
+        try {
+            final SubLevelRegionFile regionFile = this.getRegionFile(chunkPos);
+            regionFile.tryRemove(chunkPos.getRegionLocalX(), chunkPos.getRegionLocalZ());
+        } catch (final IOException e) {
+            Sable.LOGGER.error("Failed to remove holding chunk for {}", chunkPos, e);
         }
     }
 
@@ -117,9 +140,20 @@ public class SubLevelStorage implements AutoCloseable {
                 Sable.LOGGER.error("Couldn't find sub-level at index {} in storage file for chunk {}", pointer.subLevelIndex(), chunkPos);
                 return null;
             }
+
             final SubLevelData subLevel = SubLevelSerializer.fromData(tag);
 
             if (subLevel != null) {
+                final BoundingBox3d worldBounds = subLevel.bounds();
+
+                if (worldBounds.minX == 0.0 && worldBounds.minY == 0.0 && worldBounds.minZ == 0.0 &&
+                        worldBounds.maxX == 0.0 && worldBounds.maxY == 0.0 && worldBounds.maxZ == 0.0) {
+
+                    Sable.LOGGER.error("Recovering zeroed out bounds for sub-level {} loaded at {} in chunk {}", subLevel, pointer, chunkPos);
+                    final Vector3d position = subLevel.pose().position();
+                    worldBounds.set(position.x, position.y, position.z, position.x, position.y, position.z).expand(1.0);
+                }
+
                 subLevel.setOriginLoadedChunk(chunkPos);
             } else {
                 Sable.LOGGER.error("Failed to load sub-level at index {} in storage file for chunk {}", pointer.subLevelIndex(), chunkPos);
@@ -249,6 +283,41 @@ public class SubLevelStorage implements AutoCloseable {
     @ApiStatus.Internal
     public Path getFolder() {
         return this.folder;
+    }
+
+    /**
+     * Prunes all empty files in the cache
+     */
+    public void pruneCache() throws IOException {
+        final ObjectIterator<SubLevelStorageFile> storageFiles = this.storageCache.values().iterator();
+
+        while (storageFiles.hasNext()) {
+            final SubLevelStorageFile storageFile = storageFiles.next();
+
+            if (storageFile.isEmpty()) {
+                if (SableServerConfig.SUB_LEVEL_STORAGE_PRUNING.get()) {
+                    storageFile.delete();
+                } else {
+                    storageFile.close();
+                }
+                storageFiles.remove();
+            }
+        }
+
+        final ObjectIterator<SubLevelRegionFile> regionFiles = this.regionCache.values().iterator();
+
+        while (regionFiles.hasNext()) {
+            final SubLevelRegionFile regionFile = regionFiles.next();
+
+            if (regionFile.isEmpty()) {
+                if (SableServerConfig.SUB_LEVEL_STORAGE_PRUNING.get()) {
+                    regionFile.delete();
+                } else {
+                    regionFile.close();
+                }
+                regionFiles.remove();
+            }
+        }
     }
 
     /**